VIVY/requirements.txt
Flask==3.0.3
requests==2.32.3
python-dotenv==1.0.1
PyQt6==6.7.1
python-docx==1.1.2
pypdf==5.4.0
